[changed] major refactoring of the code structure
